// app/src/modules/user-roles/dtos/org-user-role-assignment-response-dto.ts
export interface OrgUserRoleAssignmentResponseDTO {
    userRoleAssignmentId: number;
    userRoleId: number;
    userRoleName: string;
    userRoleDescription?: string | null;
    userRoleOrgUnitMemberPermissionEdit: boolean;
    userRoleTeamMemberPermissionEdit: boolean;
    userRoleFormPermissionCreate: boolean;
    userRoleFormPermissionRead: boolean;
    userRoleFormPermissionEdit: boolean;
    userRoleFormPermissionDelete: boolean;
    userRoleFormPermissionAnnotate: boolean;
    userRoleFormPermissionPublish: boolean;
    userRoleProcessPermissionCreate: boolean;
    userRoleProcessPermissionRead: boolean;
    userRoleProcessPermissionEdit: boolean;
    userRoleProcessPermissionDelete: boolean;
    userRoleProcessPermissionAnnotate: boolean;
    userRoleProcessPermissionPublish: boolean;
    userRoleProcessInstancePermissionCreate: boolean;
    userRoleProcessInstancePermissionRead: boolean;
    userRoleProcessInstancePermissionEdit: boolean;
    userRoleProcessInstancePermissionDelete: boolean;
    userRoleProcessInstancePermissionAnnotate: boolean;
    orgUnitMembershipId: number;
    orgUnitMembershipOrganizationalUnitId: number;
    orgUnitMembershipOrganizationalUnitName: string;
    orgUnitMembershipOrganizationalUnitParentOrgUnitId?: number | null;
    orgUnitMembershipOrganizationalUnitDepth: number;
    orgUnitMembershipUserId: string;
    orgUnitMembershipUserFirstName: string;
    orgUnitMembershipUserLastName: string;
    orgUnitMembershipUserFullName: string;
    orgUnitMembershipUserEmail: string;
    orgUnitMembershipUserEnabled: boolean;
    orgUnitMembershipUserVerified: boolean;
    orgUnitMembershipUserGlobalAdmin: boolean;
    orgUnitMembershipUserDeletedInIdp: boolean;
}